Keep campaign tests importable without torch.
